refactor: remove ViewModelImpl
